Refactor database tables
A big part of the information in the circuit_proposal table
would also be needed in a gameroom table, so to avoid duplicating that
information this commit separates the circuit_proposal tables into
two different tables: gameroom and gameroom_proposal. It also performs
changes in the related tables to support this refactoring.

Signed-off-by: Eloá Franca Verona <eloafran@bitwise.io>